requirements.txt
google-adk
mcp[cli]
python-dotenv
google-cloud-aiplatform
absl-py
sqlalchemy